move color functions
