+++ src/lib/color.js
@@ -19,3 +19,32 @@ export const rgbaObject = (r, g, b, a) => {
 export const rgbaString = (rgba) => {
   return `rgba(${rgba.r}, ${rgba.g}, ${rgba.b}, ${rgba.a})`;
 };
+
+export const darkModeColors = () => {
+  const root = document.documentElement;
+  root.style.setProperty('--base03', '#fdf6e3');
+  root.style.setProperty('--base02', '#eee8d5');
+  root.style.setProperty('--base01', '#93a1a1');
+  root.style.setProperty('--base00', '#839496');
+  root.style.setProperty('--base0', '#657b83');
+  root.style.setProperty('--base1', '#586e75');
+  root.style.setProperty('--base2', '#073642');
+  root.style.setProperty('--base3', '#002b36');
+  // switch magenta and violet
+  root.style.setProperty('--magenta', '#6c71c4');
+  root.style.setProperty('--violet', '#d33682');
+};
+export const brightModeColors = () => {
+  const root = document.documentElement;
+  root.style.setProperty('--base3', '#fdf6e3');
+  root.style.setProperty('--base2', '#eee8d5');
+  root.style.setProperty('--base1', '#93a1a1');
+  root.style.setProperty('--base0', '#839496');
+  root.style.setProperty('--base00', '#657b83');
+  root.style.setProperty('--base01', '#586e75');
+  root.style.setProperty('--base02', '#073642');
+  root.style.setProperty('--base03', '#002b36');
+  // switch magenta and violet
+  root.style.setProperty('--magenta', '#d33682');
+  root.style.setProperty('--violet', '#6c71c4');
+};
